requirements.txt
pyTelegramBotAPI==4.12.0
paramiko==3.3.1
psutil==5.9.5
requests==2.31.0
speedtest-cli==2.1.3
